KW-485/chore: dashboard service, entry stats service 통합
