--- src/main/java/com/doubleo/passservice/domain/log/service/DashboardService.java
@@ -1,8 +1,0 @@
-package com.doubleo.passservice.domain.log.service;
-
-import com.doubleo.passservice.domain.log.dto.response.HourlyEntryResponse;
-import java.util.List;
-
-public interface DashboardService {
-    List<HourlyEntryResponse> getHourlyEntryList();
-}
