+++ src/main/java/com/doubleo/passservice/domain/stats/service/StatsService.java
@@ -1,9 +1,12 @@
 package com.doubleo.passservice.domain.stats.service;
 
+import com.doubleo.passservice.domain.log.dto.response.HourlyEntryResponse;
 import com.doubleo.passservice.domain.stats.dto.response.*;
 import java.util.List;
 
 public interface StatsService {
+
+    List<HourlyEntryResponse> getHourlyEntryList();
 
     List<DailyStatsInfoListResponse> getDailyPeriodStatsList();
