+++ src/main/java/com/doubleo/passservice/domain/stats/service/StatsServiceImpl.java
@@ -1,5 +1,7 @@
 package com.doubleo.passservice.domain.stats.service;
 
+import com.doubleo.passservice.domain.log.dto.response.HourlyEntryResponse;
+import com.doubleo.passservice.domain.log.repository.BuildingEnterLogRepository;
 import com.doubleo.passservice.domain.stats.domain.EntryStatsDaily;
 import com.doubleo.passservice.domain.stats.dto.response.*;
 import com.doubleo.passservice.domain.stats.repository.EntryStatsDailyRepository;
@@ -7,12 +9,17 @@ import com.doubleo.passservice.domain.stats.repository.EntryStatsMonthlyReposito
 import com.doubleo.passservice.domain.stats.repository.EntryStatsWeeklyRepository;
 import com.doubleo.passservice.global.util.TenantValidator;
 import java.time.DayOfWeek;
+import java.time.Duration;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.time.format.TextStyle;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,7 +31,43 @@ public class StatsServiceImpl implements StatsService {
     private final EntryStatsDailyRepository entryStatsDailyRepository;
     private final EntryStatsWeeklyRepository entryStatsWeeklyRepository;
     private final EntryStatsMonthlyRepository entryStatsMonthlyRepository;
+    private final BuildingEnterLogRepository buildingEnterLogRepository;
     private final TenantValidator tenantValidator;
+    private final RedisTemplate<String, String> redisTemplate;
+
+    @Override
+    public List<HourlyEntryResponse> getHourlyEntryList() {
+        LocalDateTime now = LocalDateTime.now().withMinute(0).withSecond(0).withNano(0);
+        String tenantId = tenantValidator.getTenantId();
+
+        LocalDateTime end = now;
+        LocalDateTime start = end.minusHours(24);
+
+        List<HourlyEntryResponse> hourlyList = new ArrayList<>();
+
+        for (int i = 0; i < 24; i++) {
+            LocalDateTime hour = start.plusHours(i);
+            String hourStr = hour.format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH"));
+            String redisKey = "hourlyEntry:" + tenantId + ":" + hourStr;
+
+            String cachedCount = redisTemplate.opsForValue().get(redisKey);
+
+            int total;
+            if (cachedCount != null) {
+                total = Integer.parseInt(cachedCount);
+            } else {
+                int count = buildingEnterLogRepository.countInLogsAtHour(hour, hour.plusHours(1));
+                redisTemplate
+                        .opsForValue()
+                        .set(redisKey, String.valueOf(count), Duration.ofSeconds(10));
+                total = count;
+            }
+
+            hourlyList.add(new HourlyEntryResponse(hour.getHour(), total, hour));
+        }
+
+        return hourlyList;
+    }
 
     @Override
     public List<DailyStatsInfoListResponse> getDailyPeriodStatsList() {
